Add thread yield capability
Current the Behaviour Tree does not yield. This could lead to side
effects like starvation and lost of interrupts on critical systems.
This add the capability to the user define if engine should perform
a k_yield() to give opportunity to other threads with the same
priority to run.

This do not affect the time slice configuration and both can be
used. The time slice can be useful when a very long action may
block tree walk. To enable time slice make sure that behaviour
tree priority is greater than TIMESLICE_PRIORITY and
TIMESLICING=y.

The best walkthru performace can be achieve when
ZEPHYR_BEHAVIOUR_TREE_ALLOW_YIELD option is disabled.

Signed-off-by: Gerson Fernando Budke <gerson.budke@ossystems.com.br>